requirements.txt
Django==4.1.5
mysqlclient==2.1.1
Pillow==9.4.0
psycopg==3.1.7
pytz==2022.7
sqlparse==0.4.3
django-filter~=22.1
django-money==3.0.0
django-datetimepicker==3.14